requirements.txt
django==4.2.16
djangorestframework==3.15.2
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
pytest-xdist==3.6.1
pytest-factoryboy==2.7.0
